# specs/001-customer-service-microservice/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Customer Service API
  description: |
    Customer Service microservice manages customer and company master data, including personal/company information,
    VAT/tax data, multiple billing/shipping addresses, NDA lifecycle management, and document references.
    The service also owns user account management for customer authentication (employee accounts managed by separate Employee Service).
  version: 1.0.0
  contact:
    name: Maliev Engineering Team
    email: engineering@maliev.com
servers:
  - url: https://api.maliev.com/customers/v1
    description: Production
  - url: https://staging.maliev.com/customers/v1
    description: Staging
  - url: http://localhost:8080/customers/v1
    description: Local Development
...
security:
  - BearerAuth: []
paths:
...
  /internal-notes:
    post:
      tags: [Internal Notes]
      summary: Create internal note
      description: Creates internal note for customer or company (employee-only, hidden from customers)
      operationId: createInternalNote
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInternalNoteRequest'
      responses:
        '201':
          description: Internal note created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalNoteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - customer users cannot create internal notes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Internal Notes]
      summary: List internal notes
      description: Retrieves internal notes for specific owner (customer or company) - employee-only
      operationId: listInternalNotes
      security:
        - BearerAuth: []
      parameters:
        - name: ownerType
          in: query
          required: true
          schema:
            type: string
            enum: [Customer, Company]
        - name: ownerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Internal notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InternalNoteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - customer users cannot view internal notes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /internal-notes/{id}:
    get:
      tags: [Internal Notes]
      summary: Get internal note by ID
      description: Retrieves internal note details (employee-only)
      operationId: getInternalNote
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InternalNoteId'
      responses:
        '200':
          description: Internal note details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalNoteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Internal Notes]
      summary: Update internal note
      description: Updates internal note text (employee-only)
      operationId: updateInternalNote
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InternalNoteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInternalNoteRequest'
      responses:
        '200':
          description: Internal note updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalNoteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ConcurrencyConflict'

    delete:
      tags: [Internal Notes]
      summary: Delete internal note
      description: Deletes internal note permanently (employee-only)
      operationId: deleteInternalNote
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InternalNoteId'
      responses:
        '204':
          description: Internal note deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # Health Checks
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
  parameters:
...
    InternalNoteId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Internal note identifier
...
  schemas:
...
    CreateInternalNoteRequest:
      type: object
      required:
        - ownerType
        - ownerId
        - noteText
      properties:
        ownerType:
          type: string
          enum: [Customer, Company]
          description: Type of entity the note is attached to
        ownerId:
          type: string
          format: uuid
          description: ID of customer or company
        noteText:
          type: string
          maxLength: 5000
          description: Internal note content (employee-only, never visible to customers)
          example: Customer prefers morning calls. Negotiating volume discount for Q3.
    UpdateInternalNoteRequest:
      type: object
      required:
        - noteText
        - version
      properties:
        noteText:
          type: string
          maxLength: 5000
          description: Updated internal note content
        version:
          type: string
          description: Base64-encoded RowVersion for optimistic concurrency
    InternalNoteResponse:
      type: object
      required:
        - id
        - ownerType
        - ownerId
        - noteText
        - createdBy
        - createdAt
        - updatedAt
        - version
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        ownerType:
          type: string
          enum: [Customer, Company]
        ownerId:
          type: string
          format: uuid
        noteText:
          type: string
          description: Internal note content
          example: Customer prefers morning calls
        createdBy:
          type: string
          description: Employee identifier who created the note
          example: employee-123
        createdAt:
          type: string
          format: date-time
          example: '2025-10-08T10:30:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2025-10-08T10:30:00Z'
        version:
          type: string
          description: Base64-encoded RowVersion for optimistic concurrency
          example: AAAAAAAABkI=

    # ============================================================================
    # Common Schemas
    # ============================================================================
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          example: BadRequest
        message:
          type: string
          example: Invalid email format
        details:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Field-level validation errors
          example:
            email: [Email must be in valid format]
            phone: [Phone must be in E.164 format]
        timestamp:
          type: string
          format: date-time
          example: '2025-10-08T10:30:00Z'
...
  responses:
    BadRequest:
      description: Bad request - validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConcurrencyConflict:
      description: Concurrency conflict - version mismatch
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Conflict
            message: The record has been modified by another user. Please retrieve the latest version and retry.
            timestamp: '2025-10-08T10:30:00Z'